feedback section

+++ src/components/Slider/Slider.jsx
@@ -2,13 +2,14 @@ import React, { Component } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import SwiperCore, { Navigation } from "swiper/core";
 import { Link } from "react-router-dom";
-import { Button, Card, Modal } from "react-bootstrap";
+import { Button, Card, Modal, Alert } from "react-bootstrap";
 import "swiper/swiper.scss";
 import "swiper/components/navigation/navigation.scss";
 import normalizeSliders from "../../utils/normalizeSliders";
 import routes from "../../routes";
 import deleteCourseById from "../../utils/deleteCourseById";
 import fetchAllCoursesById from "../../utils/fetchAllCoursesById";
+import fetchAllRoomsById from "../../utils/fetchAllRoomsById";
 import "./Slider.css";
 
 class Slider extends Component {
@@ -16,6 +17,8 @@ class Slider extends Component {
     showModal: false,
     editId: "",
     courses: [],
+    // canDeleteCourse: true,
+    modalError: "",
   };
 
   async componentDidMount() {
@@ -33,22 +36,44 @@ class Slider extends Component {
     }
   }
 
-  handleClose = () => this.setState({ showModal: false, editId: "" });
+  handleClose = () =>
+    this.setState({ showModal: false, editId: "", modalError: "" });
   handleShow = (id) => this.setState({ showModal: true, editId: id });
 
   handleDeleteCourse = async () => {
     const { editId } = this.state;
+    let canDeleteCourse = true;
 
-    await deleteCourseById(editId)
-      .then((response) => console.log(response))
-      .catch((err) => console.log(err));
+    await fetchAllRoomsById(this.props.id)
+      .then((response) => {
+        response.data.forEach((room) => {
+          if (room.courseId === editId) {
+            console.log("cannotdelete");
+            canDeleteCourse = false;
+          }
+        });
+        if (canDeleteCourse) {
+          deleteCourseById(editId)
+            .then((response) => console.log(response))
+            .catch((err) => console.log(err));
 
-    this.setState((prevState) => {
-      return {
-        showModal: false,
-        courses: prevState.courses.filter((element) => element._id !== editId),
-      };
-    });
+          this.setState((prevState) => {
+            return {
+              showModal: false,
+              courses: prevState.courses.filter(
+                (element) => element._id !== editId
+              ),
+            };
+          });
+        } else {
+          this.setState({
+            modalError: "Неможливо видалити курс, він прив'язний до кімнати!",
+          });
+        }
+      })
+      .catch((err) => {
+        console.log(err);
+      });
   };
 
   render() {
@@ -62,7 +87,7 @@ class Slider extends Component {
       },
     };
     const { classN, createRoute, isMyProfile } = this.props;
-    const { courses } = this.state;
+    const { courses, modalError } = this.state;
 
     return (
       <>
@@ -91,7 +116,7 @@ class Slider extends Component {
                           className="button d-block"
                           to={routes.course + "/" + element._id}
                         >
-                          <Button variant="primary">Show Course</Button>
+                          <Button variant="primary">Показти курс</Button>
                         </Link>
 
                         {isMyProfile && (
@@ -99,7 +124,7 @@ class Slider extends Component {
                             onClick={() => this.handleShow(element._id)}
                             variant="secondary"
                           >
-                            Edit
+                            Редагувати
                           </Button>
                         )}
                       </div>
@@ -111,11 +136,11 @@ class Slider extends Component {
           </Swiper>
         ) : (
           <div>
-            <div className="my-4 text-center">No courses yet</div>
+            <div className="my-4 text-center">Немає курсів</div>
             {isMyProfile && (
               <div className="d-flex justify-content-center">
                 <Button>
-                  <Link to={createRoute}>Create</Link>
+                  <Link to={createRoute}>Створити</Link>
                 </Button>
               </div>
             )}
@@ -123,20 +148,26 @@ class Slider extends Component {
         )}
         <Modal centered show={this.state.showModal} onHide={this.handleClose}>
           <Modal.Header closeButton>
-            <Modal.Title>You want to:</Modal.Title>
+            <Modal.Title>Виберіть дію:</Modal.Title>
           </Modal.Header>
 
           <Modal.Body className="mx-auto">
-            <Link
-              to={routes.editCourse + "/" + this.state.editId}
-              className="mr-2"
-            >
-              <Button variant="secondary">Edit</Button>
-            </Link>
+            {modalError ? (
+              <Alert variant="danger">{modalError}</Alert>
+            ) : (
+              <div>
+                <Link
+                  to={routes.editCourse + "/" + this.state.editId}
+                  className="mr-2"
+                >
+                  <Button variant="secondary">Редагувати</Button>
+                </Link>
 
-            <Button variant="danger" onClick={this.handleDeleteCourse}>
-              Delete
-            </Button>
+                <Button variant="danger" onClick={this.handleDeleteCourse}>
+                  Видалити
+                </Button>
+              </div>
+            )}
           </Modal.Body>
         </Modal>
       </>
